remove 'mydomain' from the 'mydestination' in postfix
This commit fixes a misconfig of the postofix main.cf file.

At OSE, we don't actually use postfix for accepting emails. We use GSuite (Google Apps = Gapps). We got a free plan years ago (no longer available today), and we continue to use it.

We only use postfix for sending-out emails, like server alerts and phpList newsletters. Therefore, we need to make sure that 'mydomain' isn't in the 'mydestination' list, so that postfix will know that it needs to forward any @opensourceecology.org emails out to Google Mail servers