adding 'backblaze-b2' package install to backups role
This commit updates the 'maltfield.backups' ansible role to also install the 'backblaze-b2' cli tool.

On 2002-10-28, we switched from the b2 cli tool to rclone on hetzner2 because it broke (iirc the old python version hit a TLS issue, and we couldn't update because the newer version had some new python depend that wasn't available in the Cent yum repos or something)

 * https://wiki.opensourceecology.org/wiki/Maltfield_Log/2022#Fri_October_28.2C_2022

Anyway, one benefit of us migrating to Debian in hetzner3 is that we can now install the backblaze b2 cli tool directly from the official Debian apt repos.

Probably we'll stick to rclone for uploads, but the b2 cli tool is also a requirement, as it appears to be the only way to create API keys that are restricted to append-only, which is an important security restriction that prevents our servers from being able to delete old files -- thus giving us protection from ransomware

 * https://en.wikipedia.org/wiki/Append-only