removing 'backblaze-b2' cli tool from install
This commit reverts the change I made in the last commit.

I realized that it would be very stupid to manage API keys from the server. If the server had the ability to create new API keys, then it would have the ability to mint a new key that could delete all our backups!

That defeats the purpose of only putting append-only keys on the server (for ransomware protection).