fix backups role
This commit pokes a hole in the firewall that permits the user that uploads our backups to backblaze with rclone (b2user) to be able to do so.

This commit also fixes a syntax error in the backups.sh script, where I told it to use PIGZ (parallel gzip that takes advantage of multiple processors) but also gave it the -z option (regular gzip). I fixed this by removing the '-z' argument to all the `tar` commands