fix varnish and nginx
This commit fixes our varnish vcl config's syntax errors when including the site-specific purge key

This commit also improves our nginx rate-limiting, which was breaking munin which (with its defaults, not including all monitors) does 122 requests on the first page load. So we bumpted it up to allow a burst of 300, where the first 200 happen without delay and the last 100 are slowed-down. After a burst of 300 requests, users will get a 503 error, until the 120 r/m leaky bucket drains.