enable curl in php.ini
This commit adds 'php-curl' to the list of depends to install in the php role

It also updates the php.ini config to remove the curl functions from the disabled_functions list.

I did this because the latest version of MediaWiki apparently uses some library called "GuzzleHTTP", which requires either curl or allow_url_fopen

 * https://docs.guzzlephp.org/en/stable/quickstart.html

The OWASP hardening guide for PHP tells you to disable allow_url_fopen but does not tell you to disable curl, so I figured that would be the better of the two to allow

 * https://cheatsheetseries.owasp.org/cheatsheets/PHP_Configuration_Cheat_Sheet.html

Without this change, users cannot make a request as GuzzleHTTP appears to be a prereq of the ConfirmAccount extension that we use for new user registrations. If cURL is not available, then submissions at Special:RequestAccount fail with an error

  [exception] [Z6k8UdkIYao5-602-ohiTQAAAAE] /wiki/Special:RequestAccount   RuntimeException: GuzzleHttp requires cURL, the allow_url_fopen ini setting, or a custom HTTP handler.

Note that we don't want our server to have the ability to initiate requests to the Internet (eg download something malicious or exfiltrate date by uploading it), so we have disabled this before. And we have firewall rules that prevent outgoing requests from the webserver's user. This helps, but having layered security is a good thing. For example, some months ago I realized that using ipset for my firewall rules had caused all of my firewall rules to get cleared! It ran that way for months before I caught it and reported it upstream

 * https://github.com/mikegleasonjr/ansible-role-firewall/issues/43

Layered Security is a good thing, but I guess we'll have to live with letting PHP apps have cURL access :(