add DirectoryIndex for apache configs of phpList site
This commit adds a "DirectoryIndex index.php index.html" line to the apache config for the phplist vhost.

This is required because phpList ships with both index.html and index.php files. This worked fine on hetner2, but on hetnzer3 (using php-fpm and Debian 12), the default displays the index.html as priority over index.php, which breaks some pages (eg http://phplist.opensourceecology.org/lists/?p=subscribe&id=3).

This fixes that so that it'll always prefer the .php over the .html