added deps for MediaWiki FancyCaptchca
 * https://www.mediawiki.org/wiki/Extension:ConfirmEdit#FancyCaptcha

This commit updates our php role to install necessary apt dependencies for generating self-hosted Captcha images for the ConfirmEdit and ConfirmAccount extensions, as needed for the Special:RequestAccount page.

On hetnzer2 we used SimpleCaptcha, but the latest version of SimpleCaptcha (included in ConfirmEdit and used by ConfirmAccount) appears to not work on the latest version of MediaWiki because it is missing an extension.json file.

I tried hCaptcha, but there was an issue because the server's firewall prevents the server from initiating a request to hCaptchas servers, causing timeout issues.

It's important that OSE's infrastructure is built like a "well oiled machine". OSE doesn't have a sysadmin on staff (even part-time), so it's really better to go with a self-hosted CAPTCHA system, anyway. Even if it's less effective.

Hopefully in the future they'll integrate mCaptcha. For now, FancyCapthca appears to be our only self-hosted option. It's better than nothing. And it's probably better than SimpleCaptcha. We'll see.

Anyway, FancyCaptcha needs some local font (we chose dejavu, which was installed by default in Debian 12 -- or otherwise installed as a dep by some other package that we installed). It also needs the python pillow package (not sure why MediaWiki chooses to use python instead of something native to php/pear, but that's what it requries), and it requires a word list. We chose the US wordlist, which has over 100,000 words.