execute medaiwiki cron script as root (not www-data)
This commit changes the mediawiki cron script to execute as root. Previously it was set to run as 'www-data'

This will fix the cron from failing to run. Here's what we saw in the cron's logs last night:

  root@hetzner3 /usr/local/bin # cat /var/log/cron/mediawiki_generate_captchas.log
  www-data is not in the sudoers file.
  This incident has been reported to the administrator.
  www-data is not in the sudoers file.
  This incident has been reported to the administrator.
  root@hetzner3 /usr/local/bin #

So basically the problem is that I tried to use layered security and de-escalate the privileges of this script in two places:

 1. In the script itself (with sudo)
 2. In the cron config

The most important place to de-escalate is in the script itself. That way if someone (even me) tries to run this script manually (while logged-in as a user like root or maltfield), then it would prevent the risk of arbitrary code execution or an attacker who managed to modify the contents of captcha.py (which lives in our web server root dir) from executing a privilege escalation attack.

See the note I just added to the script itself

> ###########################
> # (!!) SECURITY NOTE (!!) #
> ###########################
> # It's critical that any scripts executed inside our web server's document root
> # are executed as a non-privileged user! This is because, if there's a
> # vulnerability in MediaWiki (or wordpress, phpList, etc), then an attacker
> # can modify the contents of such scripts (eg captcha.py). And if we execute
> # that script as root, then it would give the attacker the ability a way to
> # execute arbitrary code as root, escalating their privileges.
> #
> # Also note that this script itself is executed as root by cron. It's better to
> # do the de-escalation in this script, because a human might execute it as
> # root. Also note this script wouldn't work if run as 'www-data', because
> # 'www-data' isn't in sudoers.
> #
> # Therefore, make sure any executions of scripts in /var/www/html/*/htdocs are
> # prefixed with `sudo -u www-data -bash -c "<path to script>"`
> ###########################

One solution would be to add 'www-data' to sudoers (note that we're trying to use sudo to de-escalate, not escalate), but that sounds dangerous and unnecessary.

So we'll just execute this wrapper script as root, and make sure the logic to de-escalate lives in the script itself. Hence the note to future editors of the script that it may be called as root (and should protect against the risks that come with that directly in the script itself)