re-enable blocking wp-login.php
This commit uncomments the apache config blog that causes apache to return a 403 Forbidden error when a user attempts to load a URL with 'wp-config.php' in it.

I had only temporarily done this because I couldn't login to the wordpress sites because the plugin that we had been using 'wp-rename-login' was removed from Wordpress due to security issues.

Now that I've figured out how to replace the renaming of the URL for logins with the plugin 'melapress-login-security' and configure it with the cli + mysql, we should re-enable this security feature

 * https://wiki.opensourceecology.org/wiki/Wordpress#MelaPress_Login_Security
 * https://wiki.opensourceecology.org/wiki/CHG-2025-XX-XX_migrate_store_to_hetzner3