also block /wp-content/tmp/*.php
This commit updates our apache rules to do two things for /wp-content/tmp/

 1. Never execute php in this dir

 2. Return a 403 Forbidden for any files with php extensions in this dir

This is important because this is one of several other dirs that the webserver has write-access-to.

So this hardening will prevent users from uploading files with code, and then getting the web server to execute that code (arbitrary code execution).

I tested and confirmed that this was working for all the other dirs, but this one server-writeable dir was missing.