fix munin for smart and varnish
This commit fixes our munin config for varnish and SMART disk info

I had a "managed by varnish" comment in the varnish5_ script as a comment, but it appeared before the shebang. As a result, munin was choking on it (processing it as shell instead of perl).

This also tries to add smart_, but it doesn't work because that plugin apparently dones't understand the output of NVMEs